refactor: moved userdataDisplay to pageSections
