+++ IndoorMapsFrontend/src/components/pageSections/UserDataDisplay.tsx
